requirements.txt
amqp
asgiref
async-timeout
billiard
celery
certifi
cffi
charset-normalizer
channels
channels_redis
click
click-didyoumean
click-plugins
click-repl
cron-descriptor
cryptography
Django
django-celery-beat
django-celery-results
django-cors-headers
django-redis
django-timezone-field
djangorestframework
dj-database-url
drf-yasg
drf-spectacular
gunicorn
idna
kombu
pillow
prompt_toolkit
psycopg2
pycparser
python-dotenv
PyJWT
python-crontab
python-dateutil
pyotp
pytz
redis
requests
six
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3
vine
wcwidth